[67] refactor: move menu queries to separate file
