+++ api/utils.py
@@ -5,7 +5,7 @@ Various utilities for interacting with data in API queries.
 import re
 from typing import List
 
-from .db_queries import QUERY_ONE_COLLECTION
+from queries.menu import QUERY_ONE_COLLECTION
 from .db_connection import get_db
 
 COLLECTION_PATTERN = r"^(pli-tv-b[ui]-vb|[A-Z]+[0-9]+|[a-z\-]+)"
